+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 * Wire up with API ability to assign/unassign user roles in UserForm. Refs UIU-3124.
 * Fix filter issues in user roles modal. Refs UIU-3124.
 * Suppress Settings > Users > Permission sets when `roles` interface is present. Refs UIU-3105.
+* Revert conditional use of the `users-keycloak` interface in UserRecordContainer. Refs UIU-3102.
 
 ## [10.0.4](https://github.com/folio-org/ui-users/tree/v10.0.4) (2023-11-10)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v10.0.3...v10.0.4)